+++ BestStudentCafedra/Models/ActivityProtection.cs
@@ -11,7 +11,7 @@ namespace BestStudentCafedra.Models
         public int StudentId { get; set; }
         public int ActivityId { get; set; }
         public DateTime ProtectionDate { get; set; }
-        public int Points { get; set; }
+        public float Points { get; set; }
 
         public virtual Activity Activity { get; set; }
         public virtual Student Student { get; set; }
